requirements.txt
langchain
openai
python-dotenv
streamlit>=1.8.1
numpy
json
os

